TaskGenGPT/requirements.txt
python-dotenv
jinja2
pydantic

langchain
langchain-openai
langchainhub
langsmith
langgraph